Update and add tests for Kubo 0.39+ features
- Add KuboInteropTest (4 bidirectional interop tests)
- Add Kubo040FeaturesTest
- Add Bitswap12Test
- Update all existing tests for new API signatures
- Add test.runsettings configuration

+++ test/UnixFileSystem/FileSystemNodeTest.cs
@@ -23,12 +23,12 @@ namespace Ipfs.Engine.UnixFileSystem
             };
             var link = node.ToLink("foo");
             Assert.AreEqual(node.Id, link.Id);
-            Assert.AreEqual(node.DagSize, link.Size);
+            Assert.AreEqual((ulong)node.DagSize, link.Size);
             Assert.AreEqual("foo", link.Name);
 
             link = node.ToLink();
             Assert.AreEqual(node.Id, link.Id);
-            Assert.AreEqual(node.DagSize, link.Size);
+            Assert.AreEqual((ulong)node.DagSize, link.Size);
             Assert.AreEqual("bar", link.Name);
         }
     }
